[Flyout System] Clean up storybooks
